Created projects page
I made a page to include all the projects I'm working on, with links
to their GitHub repos or whatever they need to be linked to. A lot of
CSS styling, and a lot of image formatiting to make sure things worked
correctly.

I don't know the entire changelog of these pages at this point, I need
to make a point to commit more often to log more discernable changes.