Made top navigation responsive
Two major changes.

The MAIN change is that all the top navigation headers collapse into a button, which then uses JavaScript to change the class and open/close the navigation. Mainly used for RWD, but may implement it across the whole site at a later date.

Also moved the index images to the root folder instead of having them in itis/images.